requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
dbt-bigquery>=1.7.0
google-cloud-bigquery>=3.10.0
pyarrow>=14.0.0
openpyxl>=3.1.0
